requirements.txt
boto3==1.20.52
discord.py==2.0.1
numpy==1.20.3
pandas==1.3.4
python-dotenv==0.21.0
requests==2.26.0
